flask_backend/requirements.txt
flask
python-dotenv
mysql-connector-python
python-keycloak
